Almost There!
This version is much cleaner and nicer than previous version. The only
additional major change that should be made is that the "Find Nearest
Food" function (and the associated direction calculation) should be
updated to take into account the toroidal nature of the map.

I'm guessing this will have a deglegable effect.

+++ NNTest/NNTest/Util.cs
@@ -5,16 +5,21 @@ using System.Text;
 
 namespace NNTest
 {
+    /* This utility class provide functions which are generic enough that the deserve
+     * separation from the classes in which they are called. It's possible that they will be
+     * used in many places in the code.
+     */
+
     public class Util
     {
         //Random number generator class which other classes can use
         public static Random randNumGen = new Random();
 
         //Moving average function
-        public static double CalculateMovingAverage(double oldAverage, double newValue, int windowSize)
+        public static double CalculateMovingAverage(double oldAverage, double newValue, double windowSize)
         {
-            return (oldAverage * (((double)windowSize - 1.0) / (double)windowSize)) + (newValue * (1.0 / (double)windowSize));
-
+            return (oldAverage * ((windowSize - 1.0) / windowSize)) + 
+                    (newValue * (1.0 / windowSize));
         }
     }
 }
